Test ANC-RT dataset.

A simulated dataset containing information on the number of clients attending antenatal care (ANC), the number for which their HIV status was ascertained, the number of known positive, the number who tested positive, the number who tested negative for eahc facility and time period.

ancrt

Format

A data frame with 20,000 rows and 10 variables:

faciluid

the unique facility identifier

time

the time period

n_clients

the number of clients attending antenatal care (ANC)

n_status

the number of women whose HIV status was ascertained

knownpos

the number of women who were already diangosed as living with HIV

testpos

the number of women who tested positive to HIV at ANC

testneg

the number of women who tested negative to HIV at ANC

true_prv

the true underlying HIV prevalence at this facility for the simulated dataset

snu1

the identifier of the first sub-national unit

year

the calendar year of the observation