Plot descriptive data over time

Creates plots of HIV prevalence, HIV testing coverage and, the primary data quality indicators over time

descriptive_plot(data, ylim_ind = NULL, ylim_cov = NULL,
  ylim_prv = NULL, plot_type = "full")

Arguments

data
The ANC-RT dataset. The functions check_data and data_clean should have been run on the data to properly prepare the data for use here. The dataset must include the following variables:

n_clients: The number of women who attended the specific facility during the specific time period for their first ANC visit

n_status: The number of women who attended the specific facility during the specific time period for their first ANC visit and had their HIV status ascertained

n_status_c: Cleaned n_status (generated using the data_clean function)

n_status_c_impute: n_status_c adjusted for multiple testing using the impute adjustment option (generated using the data_clean function)

n_status_c_remove: n_status_c adjusted for multiple testing using the remove adjustment option (generated using the data_clean function)

n_status_c_setmax: n_status_c adjusted for multiple testing using the set to maximum adjustment option (generated using the data_clean function)

testpos: The number of women who tested positive for HIV at their first ANC visit at the specific facility during the specific time period

testneg: The number of women who tested negative for HIV at their first ANC visit at the specific facility during the specific time period

knownpos: The number of women who attended their first ANC visit at the specific facility during the specific time period with previous knowledge of being HIV positive

totpos: Total number of positive HIV cases

totpos_c: Cleaned totpos (generated using the data_clean function)

totpos_c_impute: Adjusted totpos_c if the impute adjustment option for multiple testing is used (generated using the data_clean function)

totpos_c_remove: Adjusted totpos_c if the remove adjustment option for multiple testing is used (generated using the data_clean function)

totpos_c_setmax: Adjusted totpos_c if the set to maximum adjustment option for multiple testing is used (generated using the data_clean function)

time: The time period

ylim_ind
The y-axis lower and upper limits of the plot of the primary data quality indicators over time (if left blank, ggplot() defaults are used)

ylim_cov
The y-axis lower and upper limits of the plot of HIV testing coverage over time (if left blank, ggplot() defaults are used)

ylim_prv
The y-axis lower and upper limits of the plot of HIV prevalence over time (if left blank, ggplot() defaults are used)

plot_type
The plots included in the final plot. Options include:

full All three plots are included (default)

prev_cov Plots of HIV prevalence over time and HIV testing coverage over time are included

prev_ind Plots of HIV prevalence over time and the primary data quality indicators over time are included

cov_ind Plots of HIV testing coverage over time and the primary data quality indicators over time are included

prev The plot of HIV prevalence over time only is included

cov The plot of HIV testing coverage over time only is included

ind The plot of the primary data quality indicators over time only is included

Value

Plot(s) showing the desired descriptive data comparing the raw data, the cleaned data and the three different adjustment options for multiple testing.

Details

This function has been developed to create three plots:

The two primary data quality indicators over time. The two primary data quality indicators include: 1) The percent of observations with missing data for n_clients, n_status_c or totpos_c (in the cleaned data); and, 2) The percent of observations with invalid values for n_clients, n_status, testpos or knownpos (in the raw data).

HIV testing coverage over time. For comparison purposes, HIV testing coverage is calculated using the raw and cleaned data and also using the three adjustment options for multiple testing (1. Using the impute option; 2. Using the remove option; and, 3. Using the set to maximum option).

HIV prevalence over time. For comparison purposes, HIV prevalence is calculated using the raw and cleaned data and also using the three adjustment options for multiple testing (1. Using the impute option; 2. Using the remove option; and, 3. Using the set to maximum option).