Adjustment for missing reporting periods

Adjusts HIV prevalence and HIV testing coverage from ANC-RT data for missing data due to facilities missing reporting periods. This function has been developed to adjust HIV prevalence and HIV testing coverage from ANC-RT data for missing reporting periods within a facility. Inverse probability of censoring weighting is used to account for the missing reporting periods. The weights are calculated conditional on the facility. The adjustment is only made for missing reporting periods in between the first and last reported period in a facility to avoid adjusting for periods when the facility may have been inoperational. Specifically, the weights are calculated as followed: $$ w = \frac{1}{P(\text{Not being censored } | \text{ facility})}$$ The adjusted HIV prevalence and HIV testing coverage can be reported stratified by reporting period, year and/or subnational unit 1, according to user inputs.

hiv_prv_ipcw(data, by_period = FALSE, by_snu1 = FALSE,
  by_year = FALSE)

Arguments

data
The ANC-RT dataset. The functions check_data, data_clean and mt_adjust should have been run on the data to properly prepare the data for use here. The dataset must have the following variables:

faciluid: Facility ID.

time: The time period over which the data was collected.

n_clients: The number of women from the specified facility, during the specified time period, that attended their first ANC visit.

n_status: The number of women from the specified facility, during the specified time period, that had their HIV status ascertained at their first ANC visit, either by testing or through previous knowledge.

testpos: The number of women from the specified facility, during the specified time period, that tested positive for HIV at their first ANC visit.

knownpos: The number of women from the specified facility, during the specified time period, that already knew that they were HIV-positive at their first ANC visit.

testneg: The number of women from the specified facility, during the specified time period, that tested negative for HIV at their first ANC visit.

totpos: Total number of positive HIV cases.

totpos_c: Cleaned totpos (generated using the data_clean and mt_adjust functions).

n_status_c: Cleaned n_status (generated using the data_clean and mt_adjust functions).

snu1: The subnational unit 1 (only required if results are to be stratified by snu1).

Year: The year that the data was collected (only required if results are to be stratified by year).

by_period
TRUE indicates that the results be stratified by reporting period. FALSE indicates that the results not be stratified.

by_snu1
TRUE indicates that the results be stratified by snu1. FALSE indicates that the results not be stratified.

by_year
TRUE indicates that the results be stratified by year. FALSE indicates that the results not be stratified.

Value

A table (in dataframe format) indicating the results, stratified according to user inputs. The results table includes the following columns:

snu1 (only if results were stratified by snu1): The subnational unit 1.

time (only if results were stratified by period and/or year): The time period. When results are stratified by year and reporting period, the digits "99" are suffixed to the year (i.e. "201599" refers to the yearly result for 2015).

hiv_prv: The HIV prevalence adjusted for missing reporting periods and all previous adjustments (i.e. data cleaning and adjustment for multiple testing, if performed).

hiv_cov: The HIV testing coverage adjusted for missing reporting periods and all previous adjustments (i.e. data cleaning and adjustment for multiple testing, if performed).