Plot of HIV prevalence over time

Plots the raw and/or fully adjusted HIV prevalence over time

plot_rawadjusted(data, snu1 = "all", time_unit = "period",
  hiv_raw = TRUE, y_lim = 40)

Arguments

data
The dataframe output by impcov_adjust. The following variables must be present:

snu1: The subnational unit 1. If plots are to be generated for the different snu1s, the parameter by_snu1 = "TRUE" must have been specified when running impcov_adjust.

time: The year and/or period (i.e., quarters) that the data was collected.If plots are to be generated by reporting period, the parameter by_period = TRUE must have been specified when running impcov_adjust. If plots are to be generated by year, the parameter by_year = TRUE must have been specified when running impcov_adjust.

adjusted_prv: The HIV prevalence adjusted for imperfect testing coverage and all previous adjustments (i.e. data cleaning, adjustment for multiple testing and/or adjustment for missing reporting periods, if performed) (generated using the impcov_adjust function).

hiv_raw: The HIV prevalence using the raw data (generated using the impcov_adjust function).

snu1
The snu1 for which the plot is generated. "all" indicates that the plot include the entire country data (i.e. not snu1-specific). Default = "all".

time_unit
"period" indicates that the results be plotted by reporting period. "year" indicates that the results be plotted by year. Default = "period".

hiv_raw
"TRUE" indicates that both the fully adjusted and raw HIV prevalences be plotted. "FALSE" indicates that only the fully adjusted HIV prevalences be plotted. Default = "TRUE".

y_lim
The y-axis upper limit. Default = 40.

Value

A plot of HIV prevalence over time.

Details

This function has been developed to plot HIV prevalence over time. Both the raw and fully adjusted HIV prevalence (obtained using the data_clean, mt_adjust and impcov_adjust functions) can be plotted and compared or just the fully adjusted prevalence can be plotted. The time unit can be set to either yearly or by reporting period.