Data Quality Indicators

Calculates the different data quality indicators regarding missing data and invalid data from ANC-RT data.

quality_indicators(data, by_region = FALSE, by_time = FALSE)

Arguments

data
The ANC-RT dataset. The functions check_data and data_clean should have been run on the data to properly prepare the data for use here. The data set must have the following variables:

faciluid: The facility identification code

time: The time period

snu1: The sub-national unit 1 (only necessary if results are to be stratified by region)

n_clients: The number of women who attended the specific facility during the specific time period for their first ANC visit

n_status: The number of women who attended the specific facility during the specific time period for their first ANC visit and had their HIV status ascertained

testpos: The number of women who tested positive for HIV at their first ANC visit at the specific facility during the specific time period

testneg: The number of women who tested negative for HIV at their first ANC visit at the specific facility during the specific time period

knownpos: The number of women who attended their first ANC visit at the specific facility during the specific time period with previous knowledge of being HIV positive

totpos: The number of women who attended their first ANC visit at the specific facility during the specific time period who were HIV-positive

n_status_c: Cleaned n_status (generated using the data_clean function)

testpos_c: Cleaned testpos (generated using the data_clean function)

testneg_c: Cleaned testneg (generated using the data_clean function)

knownpos_c: Cleaned knownpos (generated using the data_clean function)

totpos_c: Cleaned totpos (generated using the data_clean function)

by_region
TRUE or FALSE to indicate whether the data quality indicators be calculated stratified by snu1

by_time
TRUE or FALSE to indicate whether the data quality indicators be calculated stratified by time

Value

A table (in dataframe format) with all data quality indicators calculated for the raw and cleaned data. If specified, data quality indicators are also displayed stratified by snu1 and/or time. The 16 data quality indicators include:

The number and % of facilities that don't report all quarters

The number and % of observations with missing data for n_clients

The number and % of observations with missing data for n_status

The number and % of observations with missing data for testpos

The number and % of observations with missing data for testneg

The number and % of observations with missing data for knownpos

The number and % of observations with missing data for at least one of n_clients, n_status, or totpos

The number and % of observations with n_status > n_clients

The number and % of observations with n_status < totpos

The number and % of observations with n_status < (testpos + testneg + knownpos)

The number and % of observations with negative values for n_clients

The number and % of observations with negative values for n_status

The number and % of observations with negative values for testpos

The number and % of observations with negative values for testneg

The number and % of observations with negative values for knownpos

The number and % of observations with invalid data for at least one of n_clients, n_status, testpos, or knownpos

Details

This function was designed to calculate 16 data quality indicators from ANC-RT data. The amount of missing data and invalid data for the five primary variables used to calculate HIV testing coverage and HIV prevalence (i.e. n_clients, n_status, testpos, testneg and, knownpos) are calculated and output in a dataframe. Data quality indicators are calculated from both raw and cleaned variables and can be calculated stratified by region and time, according to user inputs.